+++ src/packages/eslint-config-kiwicom/ourRules.js
@@ -53,7 +53,7 @@ module.exports = {
   'class-methods-use-this': OFF,
   complexity: OFF,
   'consistent-return': [ERROR, { treatUndefinedAsUnspecified: false }],
-  curly: [NEXT_VERSION_ERROR, 'all'],
+  curly: [ERROR, 'all'],
   'default-case': OFF,
   'dot-notation': WARN,
   eqeqeq: [WARN, 'smart'],
@@ -124,14 +124,14 @@ module.exports = {
   // Variables (http://eslint.org/docs/rules/#variables)
   'init-declarations': OFF,
   'no-delete-var': ERROR,
-  'no-label-var': NEXT_VERSION_ERROR,
+  'no-label-var': ERROR,
   'no-restricted-globals': OFF,
   'no-shadow-restricted-names': WARN,
   'no-shadow': OFF,
   'no-undef-init': OFF,
   'no-undef': ERROR,
   'no-undefined': OFF,
-  'no-unused-vars': [NEXT_VERSION_ERROR, { args: 'after-used' }],
+  'no-unused-vars': [ERROR, { args: 'after-used' }],
   'no-use-before-define': OFF,
 
   // Node.js and CommonJS (http://eslint.org/docs/rules/#nodejs-and-commonjs)
@@ -158,7 +158,7 @@ module.exports = {
   'id-length': OFF,
   'id-match': OFF,
   'line-comment-position': OFF,
-  'linebreak-style': NEXT_VERSION_ERROR,
+  'linebreak-style': ERROR,
   'lines-between-class-members': [
     WARN,
     'always',
@@ -245,7 +245,7 @@ module.exports = {
   ],
   'babel/new-cap': OFF,
   'babel/no-invalid-this': OFF,
-  'babel/no-unused-expressions': NEXT_VERSION_ERROR,
+  'babel/no-unused-expressions': ERROR,
   'babel/valid-typeof': ERROR,
 
   // flowtype (https://github.com/gajus/eslint-plugin-flowtype)
@@ -271,27 +271,27 @@ module.exports = {
   'jest/no-disabled-tests': WARN,
   'jest/no-focused-tests': ERROR,
   'jest/no-hooks': OFF, // TODO: WARN (?)
-  'jest/no-identical-title': NEXT_VERSION_ERROR,
-  'jest/no-jasmine-globals': NEXT_VERSION_ERROR,
-  'jest/no-jest-import': NEXT_VERSION_ERROR,
+  'jest/no-identical-title': ERROR,
+  'jest/no-jasmine-globals': ERROR,
+  'jest/no-jest-import': ERROR,
   'jest/no-large-snapshots': OFF,
   'jest/no-test-callback': OFF,
   'jest/no-test-prefixes': OFF,
-  'jest/no-test-return-statement': NEXT_VERSION_ERROR,
+  'jest/no-test-return-statement': ERROR,
   'jest/no-truthy-falsy': OFF,
   'jest/prefer-expect-assertions': OFF,
   'jest/prefer-inline-snapshots': OFF,
   'jest/prefer-spy-on': OFF,
   'jest/prefer-strict-equal': OFF, // TODO: WARN
-  'jest/prefer-to-be-null': NEXT_VERSION_ERROR,
-  'jest/prefer-to-be-undefined': NEXT_VERSION_ERROR,
-  'jest/prefer-to-contain': NEXT_VERSION_ERROR,
-  'jest/prefer-to-have-length': NEXT_VERSION_ERROR,
+  'jest/prefer-to-be-null': ERROR,
+  'jest/prefer-to-be-undefined': ERROR,
+  'jest/prefer-to-contain': ERROR,
+  'jest/prefer-to-have-length': ERROR,
   'jest/require-tothrow-message': OFF,
-  'jest/valid-describe': NEXT_VERSION_ERROR,
+  'jest/valid-describe': ERROR,
   'jest/valid-expect': WARN,
-  'jest/valid-expect-in-promise': NEXT_VERSION_ERROR,
-  'jest/prefer-todo': NEXT_VERSION_ERROR,
+  'jest/valid-expect-in-promise': ERROR,
+  'jest/prefer-todo': ERROR,
   'jest/prefer-called-with': WARN,
 
   // prefer-object-spread (https://github.com/bryanrsmith/eslint-plugin-prefer-object-spread)
@@ -307,7 +307,7 @@ module.exports = {
   'react/no-danger-with-children': OFF,
   'react/no-deprecated': WARN,
   'react/no-did-mount-set-state': OFF, // https://github.com/airbnb/javascript/issues/684#issuecomment-264094930
-  'react/no-did-update-set-state': NEXT_VERSION_ERROR,
+  'react/no-did-update-set-state': ERROR,
   'react/no-direct-mutation-state': WARN,
   'react/no-find-dom-node': OFF,
   'react/no-is-mounted': WARN,
@@ -325,7 +325,7 @@ module.exports = {
   'react/require-optimization': OFF,
   'react/require-render-return': OFF,
   'react/self-closing-comp': WARN,
-  'react/sort-comp': NEXT_VERSION_ERROR,
+  'react/sort-comp': ERROR,
   'react/sort-prop-types': OFF,
   'react/style-prop-object': OFF,
   'react/jsx-boolean-value': OFF,
@@ -344,15 +344,15 @@ module.exports = {
   'react/jsx-uses-vars': WARN,
 
   // React Hooks (https://www.npmjs.com/package/eslint-plugin-react-hooks)
-  'react-hooks/rules-of-hooks': NEXT_VERSION_ERROR,
+  'react-hooks/rules-of-hooks': ERROR,
 
   // React Accessibility (https://github.com/evcohen/eslint-plugin-jsx-a11y)
-  'jsx-a11y/accessible-emoji': NEXT_VERSION_ERROR,
-  'jsx-a11y/alt-text': WARN, // TODO: ERROR
-  'jsx-a11y/aria-props': NEXT_VERSION_ERROR,
-  'jsx-a11y/aria-unsupported-elements': NEXT_VERSION_ERROR,
+  'jsx-a11y/accessible-emoji': ERROR,
+  'jsx-a11y/alt-text': NEXT_VERSION_ERROR,
+  'jsx-a11y/aria-props': ERROR,
+  'jsx-a11y/aria-unsupported-elements': ERROR,
   'jsx-a11y/no-distracting-elements': [
-    NEXT_VERSION_ERROR,
+    ERROR,
     {
       elements: ['marquee', 'blink'],
     },
@@ -396,7 +396,7 @@ module.exports = {
     },
   ],
   'import/no-anonymous-default-export': [
-    NEXT_VERSION_ERROR,
+    ERROR,
     {
       allowArray: true,
       allowArrowFunction: false,
@@ -420,7 +420,7 @@ module.exports = {
   'monorepo/no-relative-import': ERROR,
 
   // Node.js (https://github.com/mysticatea/eslint-plugin-node)
-  'node/no-missing-require': NEXT_VERSION_ERROR,
+  'node/no-missing-require': ERROR,
   'node/no-deprecated-api': WARN,
 
   // Eslint comments (https://github.com/mysticatea/eslint-plugin-eslint-comments)
@@ -430,5 +430,5 @@ module.exports = {
 
   // Kiwi.com custom rules
   'kiwicom-incubator/only-nullable-fields': ERROR,
-  'kiwicom-incubator/no-invalid-flow-annotations': NEXT_VERSION_ERROR,
+  'kiwicom-incubator/no-invalid-flow-annotations': ERROR,
 };